Implemented the same changes to scenario load. Error messages now copied onto cmd prompt
