+++ src/scenario.c
@@ -19,6 +19,7 @@
  *****************************************************************************/
 
 #include <windows.h>
+#include "title.h"
 #include "addresses.h"
 #include "game.h"
 #include "interface/viewport.h"
@@ -110,7 +111,7 @@ int scenario_load(const char *path)
 					object_list_load();
 			}
 
-			object_read_and_load_entries(file);
+			uint8 load_success = object_read_and_load_entries(file);
 
 			// Read flags (16 bytes). Loads:
 			//	RCT2_ADDRESS_CURRENT_MONTH_YEAR
@@ -147,7 +148,12 @@ int scenario_load(const char *path)
 			sawyercoding_read_chunk(file, (uint8*)RCT2_ADDRESS_COMPLETED_COMPANY_VALUE);
 
 			fclose(file);
-
+			if (!load_success){
+				set_load_objects_fail_reason();
+				title_load();
+				rct2_endupdate();
+				return 0;//This never gets called
+			}
 			// Check expansion pack
 			// RCT2_CALLPROC_EBPSAFE(0x006757E6);
